components/devbox-connect/pyproject.toml
[project]
name = "devbox-connect"
version = "0.1.0"
description = "SSH tunnel manager for connecting to devbox ports"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paramiko>=3.4.0",
    "pyyaml>=6.0",
]

[project.scripts]
devbox-connect = "devbox_connect.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devbox_connect"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "standard"
